docs: only show alert & brand flat variants
